requirements.txt
tensorflow>=2.13.0
numpy>=1.23.0
opencv-python>=4.7.0
matplotlib>=3.5.0
scikit-learn>=1.2.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
kagglehub>=0.1.0
pillow>=9.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0